+++ .env.example
@@ -12,7 +12,6 @@ VITE_BACKEND_URL=
 
 VITE_TRANSAK_API_KEY=
 VITE_TRANSAK_URL=
-TRANSAK_API_SECRET=
 
 VITE_TRANSAK_URL_TERMS=https://transak.com/terms-of-service
 VITE_TRANSAK_URL_POLICY=https://transak.com/privacy-policy
