+++ src/pages/buy/index.ts
@@ -1,0 +1,1 @@
+export * from "./Buy";
