+++ src/pages/buy/types.d.ts
@@ -1,0 +1,8 @@
+export interface chain  {
+  "network-name": string;
+  network: string;
+  logo: string;
+  symbol: string;
+  prefix ?: number;
+  type: string;
+}
