+++ src/pages/index.ts
@@ -24,3 +24,4 @@ export * from "./swap/components";
 export * from "./swap/hooks";
 export * from "./swap";
 export * from "./welcome";
+export * from "./buy";
